requirements.txt
amqp==2.3.2
beautifulsoup4==4.6.1
billiard==3.5.0.4
bottlenose==1.1.8
bs4==0.0.1
celery==4.2.1
certifi==2018.4.16
chardet==3.0.4
dataclasses==0.6
Django==2.0.7
django-cors-headers==2.3.0
django-jet==1.0.7
django-rest-framework==0.1.0
django-rest-framework-docs==0.1.7
django-rest-multiple-models==2.1.1
djangorestframework==3.8.2
djongo==1.2.29
idna==2.7
jsonpickle==0.9.6
kombu==4.2.1
lxml==4.2.4
Pillow==5.2.0
pymongo==3.7.0
python-amazon-simple-product-api==2.2.11
python-dateutil==2.7.3
pytz==2018.5
requests==2.19.1
six==1.11.0
sqlparse==0.2.4
urllib3==1.23
vine==1.1.4
